requirements.txt
Django==2.2.4
djangorestframework
pytz==2019.2
sqlparse==0.3.0
django-cors-headers
psycopg2
pillow